Break control/ ↔ preferences/ include cycle
Part of a broader effort to clean up include dependencies as a step
towards a modular build.

Extracted `ConfigKey` from `preferences/configobject.h` into a new
`preferences/configkey.h` (no logic change, pure relocation).

Created `preferences/usersettings_fwd.h` containing only the forward
declarations of `ConfigObject<T>` / `ConfigValue` and the
`UserSettingsPointer` / `UserSettingsWeakPointer` typedefs.

Replaced `#include "preferences/usersettings.h"` with
`#include "preferences/usersettings_fwd.h"` in:
  - `control/control.h`
  - `control/controlobject.h`
  - `control/controlproxy.h`

Added direct `#include "preferences/usersettings.h"` to
`control/control.cpp`, `control/controlobject.cpp`, and
`control/controlproxy.cpp`, which need the full template definition.

Added direct `#include <QKeyEvent>` to `library/librarycontrol.h`,
which was relying on QKeyEvent being included transitively.
