+++ src/library/librarycontrol.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <QKeyEvent>
 #include <QObject>
 #include <memory>
